examples/django-project/pyproject.toml
[project]
name = "django-ecommerce-example"
version = "0.1.0"
description = "Example Django ecommerce site showcasing wilco integration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.0",
    "django-unfold>=0.40.0",
    "wilco",
]

[tool.uv.sources]
wilco = { path = "../..", editable = true }
